Add files via upload
assignment focusing on training a neural network model for sentiment analysis using the Sentiment140 dataset. The goal is to build a model architecture that avoids overfitting by monitoring training and validation loss, and accuracy across epochs. The final part of the script tests the model's performance, ensuring the validation loss slope meets certain criteria, and then saves the training history for grading