Add files via upload
assignment focused on time series forecasting using the "Daily Minimum Temperatures in Melbourne" dataset from 1981 to 1990. It involves building a neural network model with TensorFlow using a combination of LSTM and convolutional layers to predict future temperatures. The model's performance is evaluated based on its forecasting ability, and once a satisfactory result is achieved, the model is saved for grading.