Add files via upload
assignment focusing on text generation using neural networks. It uses TensorFlow to create a text vectorizer (TextVectorization) to process a corpus, train a model on sequences of words, and then predict the next word based on a seed text. After training, the model generates a continuation of the input text by iteratively predicting the next word and appending it to the seed sequence.